Pack the LHS of `fully-connected` and `dynamic-fully-connected` ops for one block of rows at a time, in the op itself.
We currently have separate `pack-lh` that produce packed inputs (of type `qp8`, `pqs8`, `pf16`, and `pf32`) that feed into `fully-connected` and `dynamic-fully-connected` ops. This means the entire input tensor is first (converted and) packed, stored in a temporary buffer, and only then read, usually only once, by a single `fully-connected` or `dynamic-fully-connected` op.

This is bad in terms of memory consumption (for the intermediate buffer), cache reuse (the inputs are packed, stored, then re-read later on), and parallel efficiency (synchronization point between the `pack-lh` and `fully-connected` ops).

This change folds the `pack-lh` operations into the `fully-connected` and `dynamic-fully-connected` ops consuming them. The `fully-connected` ops then compute the GEMM in strips of `mr`x`n`, where the `mr`x`k` left hand operand values are first packed in a per-thread buffer and then used immediately.

The resulting combined op is ~10% faster than the original two ops, both on a single and multiple threads.

PiperOrigin-RevId: 753924784