Swap strong alias with weak function
Existing state confuses LLVM CFI. Probably we can find a fix in CFI
https://github.com/llvm/llvm-project/issues/150070.

However, this workaround seems less confusing for humans as well,
and makes code compatible with existing compilers.

Before the patch the code preserved weak function
body even if weak function was overridden with strong function.

With the patch version we will override alias, which is more conventional.

PiperOrigin-RevId: 786135294