Refactor the `pthreadpool` to function like a finite state machine with the four states: - `idle`: the `pthreadpool` is initialized and waiting for a call to `pthreadpool_parallelize`, - `running`: the function `pthreadpool_parallelize` was called and there is work to be done, - `wrapping_up`: the function `pthreadpool_parallelize` was called, but all work has been allocated and we're just waiting for the threads to finish up, - `done`: the `pthreadpool` is no longer waiting for calls to `pthreadpool_parallelize` and all the threads may disband.
The threadpool `state` is protected by a spin-lock, and threads waiting on any particular state (or set of states) can use a `futex` or `pthread_cond_t` to wait on it.

Modelling the `pthreadpool` as a finite state machine makes it easier to reason about which combinations of thread states are possible, and which are not.

The new approach also means that the threads are no longer synchronized for each call to `pthreadpool_parallelize`, i.e. they float around waiting for work to appear and just do it when it is available. Work stealing ensures that the parallel computation completes even if not all threads participate.

Another improvement consists of maintaining the number of threads waiting on a `futex` or `pthread_cond_t`, and only signaling if that count is larger than zero, which effectively eliminates spurious `syscall`s and thus reduces overheads.

The new approach also allows us to dynamically set the maximum number of parallel threads for each call to `pthreadpool_parallelize` individually, which adds the following API function:
- `size_t pthreadpool_set_threads_count(pthreadpool_t threadpool, size_t num_threads)`: Tries to set the maximum number of threads that will be used during the next invocation of `pthreadpool_parallelize`, returns the number of threads effectively set.

Compared to the old implementation on a large number of XNNPACK benchmarks, the new `pthreadpool` is:
* Marginally faster on `x86_64` (with 8 threads),
* Significantly (ca. 20%) faster on `aarch64-linux` (with 8 threads),
* **Up to 2x faster** on `aarch64-android` (with 4 threads).

The speedup appears to be mostly due to the reduced number of `syscall`s, which seem to be more expensive on `aarch64-android` and `aarch64-linux` than on `x86_64-linux`.

PiperOrigin-RevId: 801755709