Change the approach to recruiting threads from a `pthreadpool_executor`.
Since calling the `pthreadpool_executor`'s `schedule` function can be expensive, the calling thread currently packed an array of data for the threads needing to be scheduled and, in turn, scheduled a single call to the `wake_up_threads` function that would schedule them.

This apporach, although more efficient than letting the calling thread do the scheduling, has high latency due to the additional level of indirection.

This change removes the `wake_up_threads` function and modifies `ensure_num_threads` to check if any threads with an ID higher than the current thread need to be scheduled, and schedules only the first two it finds.

A call to `ensure_num_threads` is added at the top of the `thread_main` function for each scheduled thread, thus ensuring that threads are scheduled in a cascading fashion.

If no threads are scheduled,
* The first call to `pthreadpool_parallelize` will call `ensure_num_threads(threadpool, 0)`, and the `pthreadpool_executor`'s `schedule` function once to schedule the first and second threads, then go on to do work.
* The first thread, at the top of the `thread_main` function, will then call `ensure_num_threads(threadpool, 1)` and schedule the third and fourth threads, then go on to do work.
* The second thread, etc...

One advantage of this approach is that `ensure_num_threads` will stop scheduling threads if the work has already been completed, thus avoiding overheads where we have a large number of threads, but very little work.

Since each thread schedules two more threads, the latency until the last thread is scheduled is at most `O(log(n))`.

PiperOrigin-RevId: 814693873